Extract sumDB feeder impl
